[platform][or1ksim] fix uart init
The call to uart_early_init was completely missing,
and when added it was noticed that a write to a non-supported
register was perfomed.
Also we can't TRACE in arch_early_init(), since the uart
hasn't been inited yet then.
That could be solved by buffering TX data and wait for the init
code, but lazy fixing it like this for now.